requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==1.2.0
python-multipart==0.0.9
httpx==0.24.1
python-dateutil==2.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pandas>=2.2.0
numpy>=1.26.0
pyserial>=3.5
matplotlib>=3.8.0 